+++ packages/jest-cli/src/args.ts
@@ -521,7 +521,7 @@ export const options: {[key: string]: Options} = {
   },
   seed: {
     description:
-      'Must be used with the `--randomize` flag. Specify the seed to randomize with.',
+      'Sets a seed value that can be retrieved in a tests file via `jest.getSeed()`. If this option is not specified Jest will randomly generate the value. The seed value must be between `-0x80000000` and `0x7fffffff` inclusive.',
     type: 'number',
   },
   selectProjects: {
@@ -553,6 +553,11 @@ export const options: {[key: string]: Options} = {
   },
   showConfig: {
     description: 'Print your jest config and then exits.',
+    type: 'boolean',
+  },
+  showSeed: {
+    description:
+      'Prints the seed value in the test report summary. See `--seed` for how to set this value',
     type: 'boolean',
   },
   silent: {