+++ packages/jest-console/src/__tests__/bufferedConsole.test.ts
@@ -20,7 +20,7 @@ describe('CustomConsole', () => {
   };
 
   beforeEach(() => {
-    _console = new BufferedConsole(() => null);
+    _console = new BufferedConsole();
   });
 
   describe('assert', () => {
@@ -40,7 +40,9 @@ describe('CustomConsole', () => {
       _console.assert(false);
 
       expect(stdout()).toMatch('AssertionError');
-      expect(stdout()).toMatch('false == true');
+      expect(stdout()).toMatch(
+        /false == true|The expression evaluated to a falsy value/,
+      );
     });
 
     test('log the assertion error when the assertion is falsy with another message argument', () => {