+++ packages/jest-core/src/collectHandles.ts
@@ -9,6 +9,8 @@
 
 import * as asyncHooks from 'async_hooks';
 import {promisify} from 'util';
+import * as v8 from 'v8';
+import * as vm from 'vm';
 import stripAnsi = require('strip-ansi');
 import type {Config} from '@jest/types';
 import {formatExecError} from 'jest-message-util';
@@ -44,6 +46,22 @@ const alwaysActive = () => true;
 const hasWeakRef = typeof WeakRef === 'function';
 
 const asyncSleep = promisify(setTimeout);
+
+let gcFunc: (() => void) | undefined = (globalThis as any).gc;
+function runGC() {
+  if (!gcFunc) {
+    v8.setFlagsFromString('--expose-gc');
+    gcFunc = vm.runInNewContext('gc');
+    v8.setFlagsFromString('--no-expose-gc');
+    if (!gcFunc) {
+      throw new Error(
+        'Cannot find `global.gc` function. Please run node with `--expose-gc` and report this issue in jest repo.',
+      );
+    }
+  }
+
+  gcFunc();
+}
 
 // Inspired by https://github.com/mafintosh/why-is-node-running/blob/master/index.js
 // Extracted as we want to format the result ourselves
@@ -124,6 +142,14 @@ export default function collectHandles(): HandleCollectionResult {
     // `close` callback runs. If someone finishes a test from the `close`
     // callback, we will not yet have seen the resource be destroyed here.
     await asyncSleep(100);
+
+    if (activeHandles.size > 0) {
+      // For some special objects such as `TLSWRAP`.
+      // Ref: https://github.com/facebook/jest/issues/11665
+      runGC();
+
+      await asyncSleep(0);
+    }
 
     hook.disable();