+++ packages/jest-repl/src/cli/runtime-cli.ts
@@ -106,6 +106,7 @@ export async function run(
         sourcesRelatedToTestsInChangedFiles: undefined,
       },
       filePath,
+      globalConfig,
     );
 
     for (const path of projectConfig.setupFiles) {