+++ packages/jest-runner/src/runTest.ts
@@ -196,6 +196,7 @@ async function runTestInternal(
         context.sourcesRelatedToTestsInChangedFiles,
     },
     path,
+    globalConfig,
   );
 
   let isTornDown = false;
@@ -280,6 +281,7 @@ async function runTestInternal(
 
   // if we don't have `getVmContext` on the env skip coverage
   const collectV8Coverage =
+    globalConfig.collectCoverage &&
     globalConfig.coverageProvider === 'v8' &&
     typeof environment.getVmContext === 'function';