+++ packages/jest-types/__typetests__/jest.test.ts
@@ -534,3 +534,6 @@ expectError(jest.retryTimes());
 
 expectType<typeof jest>(jest.setTimeout(6000));
 expectError(jest.setTimeout());
+
+expectType<number>(jest.getSeed());
+expectError(jest.getSeed(123));