+++ packages/test-utils/src/config.ts
@@ -47,7 +47,7 @@ const DEFAULT_GLOBAL_CONFIG: Config.GlobalConfig = {
   reporters: [],
   rootDir: '/test_root_dir/',
   runTestsByPath: false,
-  seed: undefined,
+  seed: 1234,
   silent: false,
   skipFilter: false,
   snapshotFormat: {},